better handling of temporary directory
When the user specifies a directory tmp with -t we create a temporary directory
tmpDir under tmp. It will have the prefix "go-docker-melt_" and will be removed
on successful program exit. If the user does not specify a temporary directory
os.TmpDir() will use the system temporary directory and create a temporary
directory with the prefix "go-docker-melt" there.

Tweaks:
- use single error variable in main()
- use golang log package functions everywhere

Signed-off-by: Christian Brauner <christian.brauner@mailbox.org>